app_backend/requirements.txt
django>=2.2.0,<2.3.0
djangorestframework>=3.11.1,<3.12.0

psycopg2>=2.7.5,<2.8.0

coreapi>=2.3.3,<2.4.0
pyyaml>=5.1.2,<5.2.0
django-rest-swagger>=2.1.2,<2.2.0
sentry-sdk>=0.17.1.,<0.18.0

python-memcached>=1.59.0,<1.60.0
python-dateutil>=2.8.1,<2.9.0

twilio>=6.45.1,<6.50.0
Pillow>=7.2.0,<7.3.0
uwsgi>=2.0.18,<2.1.0

django-cors-headers>=3.5.0,<3.6.0

pyfcm>=1.4.7,<1.5.0
razorpay>=1.2.0,<1.3.0
#
# flake8>=3.6.0,<3.7.0
